Add files via upload
I don't have any rights on the pictures nor the musics that have been used in this program. I repeat, this was a game i created when i was a 14 yo boy that had lot of spare time. Be gentle with me. I'd like to complete my dream of having a game that is created by me